change ai request shape; refactor workflow files
