+++ packages/ai/src/index.ts
@@ -15,7 +15,4 @@ export { llmLinker } from "./Code.js";
 
 export { type LlmId, type LlmName, MODEL_CONFIGS } from "./modelConfigs.js";
 
-// Export type only! We can't import SquiggleWorkflow.js because it depends on Node.js modules such as "fs".
-export { type SquiggleWorkflowInput } from "./workflows/SquiggleWorkflow.js";
-
 export { decodeWorkflowFromReader } from "./workflows/streaming.js";
