+++ packages/ai/src/server.ts
@@ -1,8 +1,6 @@
 // This file is server-only; it shouldn't be imported into client-side React components.
-export {
-  createSquiggleWorkflowTemplate,
-  fixSquiggleWorkflowTemplate,
-} from "./workflows/SquiggleWorkflow.js";
+export { createSquiggleWorkflowTemplate } from "./workflows/createSquiggleWorkflowTemplate.js";
+export { fixSquiggleWorkflowTemplate } from "./workflows/fixSquiggleWorkflowTemplate.js";
 
 export { Workflow } from "./workflows/Workflow.js";
