+++ packages/ai/src/workflows/registry.ts
@@ -1,8 +1,6 @@
 import { WorkflowTemplate } from "./ControlledWorkflow.js";
-import {
-  createSquiggleWorkflowTemplate,
-  fixSquiggleWorkflowTemplate,
-} from "./SquiggleWorkflow.js";
+import { createSquiggleWorkflowTemplate } from "./createSquiggleWorkflowTemplate.js";
+import { fixSquiggleWorkflowTemplate } from "./fixSquiggleWorkflowTemplate.js";
 
 export function getWorkflowTemplateByName(name: string) {
   const workflows: Record<string, WorkflowTemplate<any>> = Object.fromEntries(
