+++ packages/hub/src/app/ai/Sidebar.tsx
@@ -17,7 +17,7 @@ import {
   TextAreaFormField,
 } from "@quri/ui";
 
-import { CreateRequestBody } from "./utils";
+import { AiRequestBody } from "./utils";
 import { WorkflowSummaryList } from "./WorkflowSummaryList";
 
 type Handle = {
@@ -25,7 +25,7 @@ type Handle = {
 };
 
 type Props = {
-  submitWorkflow: (requestBody: CreateRequestBody) => void;
+  submitWorkflow: (requestBody: AiRequestBody) => void;
   selectWorkflow: (id: string) => void;
   selectedWorkflow: ClientWorkflow | undefined;
   workflows: ClientWorkflow[];
@@ -80,12 +80,19 @@ Outputs:
   }));
 
   const handleSubmit = form.handleSubmit(
-    async ({ prompt, squiggleCode, model }, event) => {
-      const requestBody: CreateRequestBody = {
-        prompt: mode === "create" ? prompt : undefined,
-        squiggleCode: mode === "edit" ? squiggleCode : undefined,
-        model: model as LlmId,
-      };
+    async ({ prompt, squiggleCode, model }) => {
+      const requestBody: AiRequestBody =
+        mode === "create"
+          ? {
+              kind: "create",
+              prompt,
+              model: model as LlmId,
+            }
+          : {
+              kind: "edit",
+              squiggleCode,
+              model: model as LlmId,
+            };
 
       submitWorkflow(requestBody);
       form.setValue("prompt", "");
